# schemas/openapi.yaml
openapi: 3.1.0
info:
  title: Agent OS Control Plane API
  description: |
    API for the Agent OS Control Plane — vendor-neutral orchestration,
    governance, and observability for AI agents.
  version: 0.1.0
  contact:
    name: Agent OS Team
    email: api@agent-os.dev
  license:
    name: Proprietary
    url: https://agent-os.dev/license
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
  - url: https://api.agent-os.dev/v1
    description: Production
security:
  - bearerAuth: []
paths:
...
  /agents:
    get:
      operationId: listAgents
      summary: List all agents
      tags: [Agents]
      parameters:
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive, error]
      responses:
        '200':
          description: Agents listed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentList'
    post:
      operationId: createAgent
      summary: Create a new agent
      tags: [Agents]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentCreate'
      responses:
        '201':
          description: Agent created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '400':
          $ref: '#/components/responses/BadRequest'
  /agents/{agent_id}:
    get:
      operationId: getAgent
      summary: Get agent by ID
      tags: [Agents]
      parameters:
        - $ref: '#/components/parameters/AgentIdParam'
      responses:
        '200':
          description: Agent found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      operationId: updateAgent
      summary: Update agent
      tags: [Agents]
      parameters:
        - $ref: '#/components/parameters/AgentIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentUpdate'
      responses:
        '200':
          description: Agent updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      operationId: deleteAgent
      summary: Delete agent
      tags: [Agents]
      parameters:
        - $ref: '#/components/parameters/AgentIdParam'
      responses:
        '204':
          description: Agent deleted
        '404':
          $ref: '#/components/responses/NotFound'
...
  /memory:
    post:
      operationId: createMemory
      summary: Add to memory
      tags: [Memory]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemoryCreate'
      responses:
        '201':
          description: Memory created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Memory'
    get:
      operationId: searchMemory
      summary: Search memory
      tags: [Memory]
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Memory results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemorySearchResults'
  /memory/{key}:
    get:
      operationId: getMemory
      summary: Get memory by key
      tags: [Memory]
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Memory found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Memory'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      operationId: deleteMemory
      summary: Delete memory
      tags: [Memory]
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Memory deleted
        '404':
          $ref: '#/components/responses/NotFound'
...
  /workspaces:
    get:
      operationId: listWorkspaces
      summary: List persisted workspaces
      tags: [Control Plane]
      responses:
        '200':
          description: Workspaces listed
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Workspace' }
    post:
      operationId: createWorkspace
      summary: Create a workspace boundary
      tags: [Control Plane]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WorkspaceCreate' }
      responses:
        '201':
          description: Workspace created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Workspace' }
  /missions:
    get:
      operationId: listMissions
      summary: List persisted missions
      tags: [Control Plane]
      parameters:
        - name: workspace_id
          in: query
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Missions listed
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Mission' }
    post:
      operationId: createMission
      summary: Create a workspace-scoped mission
      tags: [Control Plane]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MissionCreate' }
      responses:
        '201':
          description: Mission created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Mission' }
        '404': { $ref: '#/components/responses/NotFound' }
  /automations:
    get:
      operationId: listAutomations
      summary: List persisted automations
      tags: [Control Plane]
      responses:
        '200':
          description: Automations listed
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Automation' }
    post:
      operationId: createAutomation
      summary: Create a workspace-scoped automation
      tags: [Control Plane]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AutomationCreate' }
      responses:
        '201':
          description: Automation created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Automation' }
  /approvals:
    get:
      operationId: listApprovals
      summary: List approval records
      tags: [Control Plane]
      responses:
        '200':
          description: Approvals listed
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Approval' }
    post:
      operationId: createApproval
      summary: Request a durable human decision
      tags: [Control Plane]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ApprovalCreate' }
      responses:
        '201':
          description: Approval requested
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Approval' }
  /approvals/{approval_id}/decision:
    post:
      operationId: decideApproval
      summary: Record an approval decision once
      tags: [Control Plane]
      parameters:
        - name: approval_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ApprovalDecision' }
      responses:
        '200':
          description: Decision recorded
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Approval' }
        '409':
          description: Approval was already decided
  /audit-events:
    get:
      operationId: listAuditEvents
      summary: List append-only workspace audit evidence
      tags: [Control Plane]
      parameters:
        - name: workspace_id
          in: query
          required: true
          schema: { type: string, format: uuid }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
      responses:
        '200':
          description: Audit events listed
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AuditEvent' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
    OffsetParam:
      name: offset
      in: query
      schema:
        type: integer
        default: 0
        minimum: 0
    AgentIdParam:
      name: agent_id
      in: path
      required: true
      schema:
        type: string
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    WorkspaceCreate:
      type: object
      required: [name]
      properties:
        name: { type: string }
        description: { type: string, default: '' }
        budget: { type: number, minimum: 0, default: 0 }
    Workspace:
      allOf:
        - $ref: '#/components/schemas/WorkspaceCreate'
        - type: object
          required: [id, status, created_at, updated_at]
          properties:
            id: { type: string, format: uuid }
            status: { type: string, example: active }
            created_at: { type: string, format: date-time }
            updated_at: { type: string, format: date-time }
    MissionCreate:
      type: object
      required: [workspace_id, title, objective]
      properties:
        workspace_id: { type: string, format: uuid }
        title: { type: string }
        objective: { type: string }
        plan: { type: array, items: { type: object } }
    Mission:
      allOf:
        - $ref: '#/components/schemas/MissionCreate'
        - type: object
          required: [id, status, progress, evidence, created_at, updated_at]
          properties:
            id: { type: string, format: uuid }
            status: { type: string }
            progress: { type: integer, minimum: 0, maximum: 100 }
            evidence: { type: array, items: { type: object } }
            created_at: { type: string, format: date-time }
            updated_at: { type: string, format: date-time }
    AutomationCreate:
      type: object
      required: [workspace_id, name, trigger_type]
      properties:
        workspace_id: { type: string, format: uuid }
        name: { type: string }
        description: { type: string }
        trigger_type: { type: string }
        trigger_config: { type: object }
        steps: { type: array, items: { type: object } }
    Automation:
      allOf:
        - $ref: '#/components/schemas/AutomationCreate'
        - type: object
          required: [id, enabled, created_at, updated_at]
          properties:
            id: { type: string, format: uuid }
            enabled: { type: integer, enum: [0, 1] }
            created_at: { type: string, format: date-time }
            updated_at: { type: string, format: date-time }
    ApprovalCreate:
      type: object
      required: [mission_id, action]
      properties:
        mission_id: { type: string, format: uuid }
        action: { type: string }
        scope: { type: object }
    ApprovalDecision:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [approved, rejected] }
        decision_note: { type: string }
    Approval:
      allOf:
        - $ref: '#/components/schemas/ApprovalCreate'
        - type: object
          required: [id, status, created_at]
          properties:
            id: { type: string, format: uuid }
            status: { type: string, enum: [pending, approved, rejected] }
            decision_note: { type: [string, 'null'] }
            created_at: { type: string, format: date-time }
            decided_at: { type: [string, 'null'], format: date-time }
    AuditEvent:
      type: object
      required: [id, workspace_id, event_type, resource_type, resource_id, actor, details, created_at]
      properties:
        id: { type: string, format: uuid }
        workspace_id: { type: string, format: uuid }
        event_type: { type: string, example: approval.approved }
        resource_type: { type: string, example: approval }
        resource_id: { type: string, format: uuid }
        actor: { type: string, example: local-system }
        details: { type: object }
        created_at: { type: string, format: date-time }
...
    Agent:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        model:
          type: string
        status:
          type: string
          enum: [active, inactive, error]
        capabilities:
          type: array
          items:
            type: string
        config:
          type: object
        policies:
          type: object
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    AgentCreate:
      type: object
      required: [name, model]
      properties:
        name:
          type: string
        model:
          type: string
        capabilities:
          type: array
          items:
            type: string
        config:
          type: object
        policies:
          type: object
    AgentUpdate:
      type: object
      properties:
        name:
          type: string
        model:
          type: string
        status:
          type: string
          enum: [active, inactive]
        config:
          type: object
        policies:
          type: object
    AgentList:
      type: object
      properties:
        agents:
          type: array
          items:
            $ref: '#/components/schemas/Agent'
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
...
    Memory:
      type: object
      properties:
        key:
          type: string
        content:
          type: string
        metadata:
          type: object
        created_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
    MemoryCreate:
      type: object
      required: [key, content]
      properties:
        key:
          type: string
        content:
          type: string
        metadata:
          type: object
        ttl:
          type: integer
          description: Time to live in seconds
    MemorySearchResults:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/Memory'
        total:
          type: integer
...
    Error:
      type: object
      properties:
        detail:
          type: string
        code:
          type: string